# packages/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "CheckoutBillingDto": {
        "properties": {
          "addressLine1": {
            "maxLength": 200,
            "minLength": 1,
            "pattern": "\\S",
            "type": "string"
          },
          "addressLine2": {
            "maxLength": 200,
            "type": "string"
          },
          "city": {
            "maxLength": 100,
            "minLength": 1,
            "pattern": "\\S",
            "type": "string"
          },
          "companyId": {
            "maxLength": 50,
            "type": "string"
          },
          "companyName": {
            "maxLength": 200,
            "type": "string"
          },
          "countryCode": {
            "maxLength": 2,
            "minLength": 2,
            "pattern": "^[A-Z]{2}$",
            "type": "string"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "pattern": "\\S",
            "type": "string"
          },
          "postalCode": {
            "maxLength": 20,
            "minLength": 1,
            "pattern": "\\S",
            "type": "string"
          },
          "vatId": {
            "maxLength": 50,
            "type": "string"
          }
        },
        "required": [
          "name",
          "addressLine1",
          "city",
          "postalCode",
          "countryCode"
        ],
        "type": "object"
      },
...
      "CheckoutPaymentDto": {
        "properties": {
          "amountMinor": {
            "maximum": 9007199254740991,
            "minimum": 1,
            "type": "integer"
          },
          "checkoutUrl": {
            "format": "uri",
            "nullable": true,
            "type": "string"
          },
          "currency": {
            "maxLength": 3,
            "minLength": 3,
            "type": "string"
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string"
          },
          "method": {
            "enum": [
              "CARD",
              "BANK_TRANSFER"
            ],
            "type": "string"
          },
          "paymentId": {
            "format": "uuid",
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "status": {
            "enum": [
              "CREATED",
              "PENDING",
              "CAPTURED",
              "FAILED",
              "VOIDED",
              "REFUND_PENDING",
              "PARTIALLY_REFUNDED",
              "REFUNDED"
            ],
            "type": "string"
          }
        },
        "required": [
          "paymentId",
          "provider",
          "method",
          "status",
          "amountMinor",
          "currency",
          "checkoutUrl",
          "expiresAt"
        ],
        "type": "object"
      },
...
      "CreateBalancePaymentDto": {
        "properties": {
          "method": {
            "enum": [
              "CARD",
              "BANK_TRANSFER"
            ],
            "type": "string"
          }
        },
        "required": [
          "method"
        ],
        "type": "object"
      },
      "CreateCheckoutPaymentDto": {
        "properties": {
          "acceptClaimPolicy": {
            "type": "boolean"
          },
          "acceptTerms": {
            "type": "boolean"
          },
          "acknowledgeWithdrawalException": {
            "type": "boolean"
          },
          "billing": {
            "$ref": "#/components/schemas/CheckoutBillingDto"
          },
          "claimPolicyRevision": {
            "maxLength": 100,
            "minLength": 1,
            "pattern": "\\S",
            "type": "string"
          },
          "email": {
            "format": "email",
            "maxLength": 320,
            "type": "string"
          },
          "fullName": {
            "maxLength": 200,
            "minLength": 1,
            "pattern": "\\S",
            "type": "string"
          },
          "method": {
            "enum": [
              "CARD",
              "BANK_TRANSFER"
            ],
            "type": "string"
          },
          "photoConsentRevision": {
            "maxLength": 100,
            "nullable": true,
            "type": "string"
          },
          "photoPublicationConsent": {
            "type": "boolean"
          },
          "termsRevision": {
            "maxLength": 100,
            "minLength": 1,
            "pattern": "\\S",
            "type": "string"
          }
        },
        "required": [
          "email",
          "fullName",
          "billing",
          "method",
          "acceptTerms",
          "acceptClaimPolicy",
          "termsRevision",
          "claimPolicyRevision",
          "acknowledgeWithdrawalException",
          "photoPublicationConsent"
        ],
        "type": "object"
      },
...
      "PaymentWebhookAcceptedDto": {
        "properties": {
          "outcome": {
            "type": "string"
          }
        },
        "required": [
          "outcome"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "opaque",
        "scheme": "bearer",
        "type": "http"
      },
      "operatorSession": {
        "description": "Environment-specific HttpOnly operator session cookie; development uses a local cookie name.",
        "in": "cookie",
        "name": "__Host-taven_admin",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Canonical HTTP contract for Taven clients",
    "title": "Taven API",
    "version": "0.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/orders/{orderId}/balance-payment": {
      "post": {
        "operationId": "PaymentsController_createBalance",
        "parameters": [
          {
            "description": "Stable client command identity (8-255 characters)",
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "minLength": 8,
              "pattern": "^\\s*\\S[\\s\\S]{6,253}\\S\\s*$",
              "type": "string"
            }
          },
          {
            "description": "Required for unsafe operator requests. Obtain the session-bound value from GET /admin/auth/session.",
            "in": "header",
            "name": "x-csrf-token",
            "required": true,
            "schema": {
              "maxLength": 256,
              "minLength": 20,
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "orderId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBalancePaymentDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutPaymentDto"
                }
              }
            },
            "description": ""
          },
          "409": {
            "description": "Balance topology or command changed"
          },
          "503": {
            "description": "The configured payment provider is unavailable"
          }
        },
        "security": [
          {
            "operatorSession": []
          }
        ],
        "summary": "Create the post-QC balance payment intent",
        "tags": [
          "payments"
        ]
      }
    },
...
    "/automatic-quote-sessions/{sessionId}/checkout/payment": {
      "delete": {
        "operationId": "PaymentsController_cancel",
        "parameters": [
          {
            "in": "query",
            "name": "paymentId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "sessionId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutPaymentDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancel an open checkout payment",
        "tags": [
          "payments"
        ]
      },
      "get": {
        "operationId": "PaymentsController_status",
        "parameters": [
          {
            "in": "query",
            "name": "paymentId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "sessionId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutPaymentDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Read a specific checkout payment state",
        "tags": [
          "payments"
        ]
      }
    },
    "/automatic-quote-sessions/{sessionId}/checkout/payments": {
      "post": {
        "operationId": "PaymentsController_create",
        "parameters": [
          {
            "description": "Stable client command identity (8-255 characters)",
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "minLength": 8,
              "pattern": "^\\s*\\S[\\s\\S]{6,253}\\S\\s*$",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "sessionId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCheckoutPaymentDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutPaymentDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "description": "Session capability is invalid"
          },
          "409": {
            "description": "Checkout topology or input changed"
          },
          "503": {
            "description": "Checkout awaits launch approval or provider recovery"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create the initial checkout payment intent",
        "tags": [
          "payments"
        ]
      }
    },
...
    "/payments/webhooks/{provider}": {
      "post": {
        "operationId": "PaymentsController_webhook",
        "parameters": [
          {
            "in": "path",
            "name": "provider",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "additionalProperties": true,
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentWebhookAcceptedDto"
                }
              }
            },
            "description": ""
          },
          "429": {
            "description": "Payment webhook verification limit is exhausted"
          }
        },
        "summary": "Consume an authenticated payment-provider event",
        "tags": [
          "payments"
        ]
      }
    },
...
  },
  "servers": [],
...
}